pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "inngest>=0.5.6",
    "llama-index-core>=0.14.0",
    "llama-index-readers-file>=0.5.4",
    "openai>=1.107.0",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "streamlit>=1.49.1",
    "uvicorn>=0.35.0",
    "groq >= 0.9.0",
]
